automatically backed up by learn

+++ lib/create.sql
@@ -15,7 +15,7 @@ CREATE TABLE users (
 
 CREATE TABLE pledges (
   id INTEGER PRIMARY KEY,
-  amount INTEGER,
+  amount NUMERIC,
   user_id INTEGER,
   project_id INTEGER
 );
